backend/requirements.txt
# Web Framework & Server
fastapi[standard]>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0

# Database & Auth
supabase>=2.3.0

# AI Services
anthropic>=0.19.0

# Caching
upstash-redis>=1.0.0

# Environment & Utilities
python-dotenv>=1.0.1
httpx>=0.27.0